+++ backend/routes/socket.js
@@ -4,6 +4,15 @@ const cors = require("cors");
 const { Server } = require("socket.io");
 const http = require("http");
 const debugPrint = require("../utils/debugPrint");
+const {
+  transformToLegacyFormat,
+  activateUser,
+  buildMsg,
+  UsersState,
+  getUser,
+  userLeavesApp,
+  userLeavesRoom,
+} = require("./usersState");
 
 app.use(cors());
 const server = http.createServer(app);
@@ -16,90 +25,6 @@ const io = new Server(server, {
   maxHttpBufferSize: 2e7,
 });
 
-// adding server-side Users state
-const UsersState = {
-  users: [], // user: {id: string; username: string; rooms: string[]}
-  setUsers: function (newUsersArray) {
-    this.users = newUsersArray;
-  },
-};
-
-// User functions
-function activateUser(id, username, roomId) {
-  // finding if user present in UsersState
-  const user = getUser(id);
-
-  // default new user
-  let newUser = {};
-
-  // user already in the UsersState, updating user data
-  if (user) {
-    newUser = { ...user, rooms: [...user.rooms, roomId] };
-  } else {
-    newUser = {
-      id,
-      username,
-      rooms: [roomId],
-    };
-  }
-
-  // updating UsersState
-  UsersState.setUsers([
-    ...UsersState.users.filter((user) => user.id !== id),
-    newUser,
-  ]);
-
-  return newUser;
-}
-
-// remove user from a room
-function userLeavesRoom(id, roomId) {
-  const user = getUser(id);
-  const updatedUser = {
-    ...user,
-    rooms: user.rooms.filter((room) => room !== roomId),
-  };
-  UsersState.setUsers([
-    ...UsersState.users.filter((user) => user.id !== id),
-    updatedUser,
-  ]);
-}
-
-// remove user from UsersState
-function userLeavesApp(id) {
-  UsersState.setUsers([...UsersState.users.filter((user) => user.id !== id)]);
-}
-
-function getUser(id) {
-  return UsersState.users.find((user) => user.id === id);
-}
-
-function getUsersInRoom(roomId) {
-  return UsersState.users.filter((user) => user.rooms.includes(roomId));
-}
-
-function getAllActiveRooms() {
-  return Array.from(new Set(UsersState.users.map((user) => user.rooms).flat()));
-}
-
-function buildMsg(text, roomId) {
-  return {
-    text,
-    socketId: "kurakani",
-    roomId,
-  };
-}
-
-function transformToLegacyFormat(usersState) {
-  const allRooms = getAllActiveRooms();
-  const result = {};
-  allRooms.forEach((roomId) => {
-    result[roomId] = getUsersInRoom(roomId).map((user) => user.id);
-  });
-
-  return result;
-}
-
 io.on("connection", (socket) => {
   io.emit("users_response", transformToLegacyFormat(UsersState));
   debugPrint(`User Connected: ${socket.id}`);
@@ -111,22 +36,10 @@ io.on("connection", (socket) => {
     // joining the room
     socket.join(roomId);
 
-    // sending message to the joined user
-    socket.emit(
-      "receive_message",
-      buildMsg(
-        `You have joined the ${roomId === "1" ? "Global" : roomId} chat room`,
-        roomId
-      )
-    );
-
     // to everyone else in the room
     socket.broadcast
       .to(roomId)
-      .emit(
-        "receive_message",
-        buildMsg(`${user.username} has joined the room`, roomId)
-      );
+      .emit("receive_message", buildMsg(`${user.username} has joined`, roomId));
 
     // converting new UsersState data to legacy state format for front-end compatibility
     io.emit("users_response", transformToLegacyFormat(UsersState));
@@ -159,7 +72,7 @@ io.on("connection", (socket) => {
       user.rooms.forEach((roomId) => {
         io.emit(
           "receive_message",
-          buildMsg(`${user.username} has left the room`, roomId)
+          buildMsg(`${user.username} has left`, roomId)
         );
       });
     }
@@ -167,10 +80,7 @@ io.on("connection", (socket) => {
 
   socket.on("leave_room", ({ username, roomId }) => {
     userLeavesRoom(socket.id, roomId);
-    io.emit(
-      "receive_message",
-      buildMsg(`${username} has left the room`, roomId)
-    );
+    io.emit("receive_message", buildMsg(`${username} has left`, roomId));
   });
 });
